.travis/requirements.txt
requests
loguru
python-box
retry
